Стандартизировали константы для Player/CellularAutomatonPlayer.cs, вынесли реализацию плеера и загрузку записи из Player/CellularAutomatonPlayer.cs в PlayerController.cs, исправили мелкие ошибки

+++ CellularAutomaton.Components/Player/IPlayer.cs
@@ -53,7 +53,7 @@ namespace CellularAutomaton.Components.Player
         /// <summary>
         /// Возвращает или задаёт скорость воспроизведения (кадры в минуту).
         /// </summary>
-        byte FramesPerMinute { get; set; }
+        short FramesPerMinute { get; set; }
 
         /// <summary>
         /// Возвращает воспроизводимую запись.
